CrewAI_Apps/requirements.txt
crewai 
serpapi
pydantic 
crewai-tools 
python-dotenv 
google-search-results